+++ changelog.md
@@ -1,4 +1,9 @@
 # Better Icon Randomizer Changelog
+## v1.3.0 (2024-05-04)
+- Rewrote the codebase to be more efficient
+- Added ship fire randomization
+- Added iOS support
+
 ## v1.2.3 (2024-04-30)
 - Added color to the randomize button color setting display
 
@@ -10,6 +15,7 @@
 
 ## v1.2.0 (2024-04-14)
 - Randomized the color of the randomize button
+- Added ability to change the color of the randomize button
 - Added macOS support
 
 ## v1.1.1 (2024-04-05)